\name{check_dependencies}
\alias{check_dependencies}
\title{Check all dependencies in the current script}
\usage{
check_dependencies()
}
\description{
 This addin scans the current R script and installs any
  dependencies that the script requires but are not installed.
}
\examples{
check_dependencies()
}

