\name{install_dependencies}
\alias{install_dependencies}
\title{Install all dependencies in the current script}
\usage{
install_dependencies()
}
\description{
 This addin scans the current R script and installs any
  dependencies that the script requires but are not installed.
}
\examples{
install_dependencies()
}

